+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Enraiged Forms Release Notes
 
+2026-02-13 : 0.3.19.3  
+Update: Improved form handling and security
+
+- Added ability to assert secure on form sections and tabs
+- Added ability to directly modify template {pre,post}content,header
+- Changed logic to call form prepare method from the constructor
+- Improved handling of assembling form actions to prevent 500 errors
+
+---
+
 2026-02-05 : 0.3.19.2  
 Bugfix: Corrected issue preparing checkbox,switch field values